Updated NordLynxExtractKeys.sh to detect prior connected status.
Executes commands once and captures output. Then uses that output going forward.

Checks to see if you are already connected to nordvpn and in that case just extracts the information of the current connection.

Output format easier to change going forward.

Added 'Server Name' , 'Server City' and 'Server Country' to the output.